Moved the trimming to its own function in the relevant file
